+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Enchantment token crash with mod-added enchantment types [#394]
+
 ## [3.3.9]
 Requires Silent Gear 1.3.9 or higher (if installed)
 ### Added